+++ python/src/services/__init__.py
@@ -1,6 +1,6 @@
 """Services package."""
 
 from services.datasus_service import DatasusService
-from services.datasus import DatasusIBGEService, DatasusSIHService
+from services.datasus import DatasusCIHService, DatasusIBGEService, DatasusSIHService
 
-__all__ = ["DatasusService", "DatasusIBGEService", "DatasusSIHService"]
+__all__ = ["DatasusService", "DatasusCIHService", "DatasusIBGEService", "DatasusSIHService"]